added csg.c

+++ mathlib.h
@@ -1,7 +1,7 @@
 #pragma once
 #include <math.h> 
 
-typedef float vec_t;
+typedef double vec_t;
 typedef vec_t vec3_t[3];
 
 extern vec3_t vec3_origin;
